Adding CORS response headers to APIs
Because:
- I want to call these APIs from the web console

I changed the start and terminate APIs to follow the Lambda proxy
integration format like the List API was.  I also added the
Access-Control-Allow-Origin response header to each of them, hardcoded
to the Netlify DNS name that the console is hosted with.  I removed a
test handler as well.